feat(embedded): flesh out BLE * Moved BLE code from main into embedded/sr/nearby_ble and fleshed it out some, updating to support setting raw advertise data so that it works with the nearby ble api. * Tested and the fast pair validator app recognizes it and can perform the calibration routine with it, but of course not the fast pairing
