+++ components/embedded/include/embedded.hpp
@@ -12,6 +12,14 @@
 #include "nearby_platform_trace.h"
 #include "nearby_fp_client.h"
 
+#include <esp_bt.h>
+#include <esp_gap_ble_api.h>
+#include <esp_gatts_api.h>
+#include <esp_bt_defs.h>
+#include <esp_bt_device.h>
+#include <esp_bt_main.h>
+#include <esp_gatt_common_api.h>
+#include <esp_random.h>
 #include <nvs_flash.h>
 
 #include "logger.hpp"
